+++ src/Symfony/Bundle/FrameworkBundle/Command/AssetsInstallCommand.php
@@ -82,9 +82,7 @@ EOT
         $targetArg = rtrim($input->getArgument('target'), '/');
 
         if (!is_dir($targetArg)) {
-            $appRoot = $this->getContainer()->getParameter('kernel.root_dir').'/..';
-
-            $targetArg = $appRoot.'/'.$targetArg;
+            $targetArg = $this->getContainer()->getParameter('kernel.project_dir').'/'.$targetArg;
 
             if (!is_dir($targetArg)) {
                 throw new \InvalidArgumentException(sprintf('The target directory "%s" does not exist.', $input->getArgument('target')));